- Move BuyMeACoffee.kt to components package - Add analytics to track click impressions upon BuyMeACoffee button - Code cleanup
